refactor(zb_io): split monolothic install functionality into submodules
