--- zb_io/src/installer/install.rs
@@ -1,3339 +1,0 @@
-use std::collections::BTreeMap;
-use std::fs::{self, File};
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use tracing::warn;
-
-use fs4::fs_std::FileExt;
-
-use crate::cellar::link::Linker;
-use crate::cellar::materialize::Cellar;
-use crate::installer::cask::resolve_cask;
-use crate::network::api::ApiClient;
-use crate::network::cache::ApiCache;
-use crate::network::download::{
-    DownloadProgressCallback, DownloadRequest, DownloadResult, ParallelDownloader,
-};
-use crate::progress::{InstallProgress, ProgressCallback};
-use crate::storage::blob::BlobCache;
-use crate::storage::db::Database;
-use crate::storage::store::Store;
-
-use zb_core::{
-    BuildPlan, Error, Formula, InstallMethod, SelectedBottle, formula_token, resolve_closure,
-    select_bottle,
-};
-
-/// Maximum number of retries for corrupted downloads
-const MAX_CORRUPTION_RETRIES: usize = 3;
-
-pub struct Installer {
-    api_client: ApiClient,
-    downloader: ParallelDownloader,
-    store: Store,
-    cellar: Cellar,
-    linker: Linker,
-    db: Database,
-    prefix: std::path::PathBuf,
-    locks_dir: PathBuf,
-}
-
-#[derive(Debug)]
-pub struct PlannedInstall {
-    pub install_name: String,
-    pub formula: Formula,
-    pub method: InstallMethod,
-}
-
-#[derive(Debug)]
-pub struct InstallPlan {
-    pub items: Vec<PlannedInstall>,
-}
-
-pub struct ExecuteResult {
-    pub installed: usize,
-}
-
-/// A package that has a newer version available upstream.
-#[derive(Debug, Clone, serde::Serialize)]
-pub struct OutdatedPackage {
-    pub name: String,
-    pub installed_version: String,
-    pub current_version: String,
-    #[serde(skip)]
-    pub installed_sha256: String,
-    #[serde(skip)]
-    pub current_sha256: String,
-    /// Whether this was installed from source (vs bottle)
-    #[serde(skip)]
-    pub is_source_build: bool,
-}
-
-impl Installer {
-    // FIXME: Create a config struct for this
-    // Then we can have `Installer::new(config: InstallerConfig, api_client: ApiClient, ...)`
-    // and derive `root.join("locks")` from config.root
-    // This will eventually remove locks_dir and keep root/prefix in one place
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        api_client: ApiClient,
-        blob_cache: BlobCache,
-        store: Store,
-        cellar: Cellar,
-        linker: Linker,
-        db: Database,
-        prefix: std::path::PathBuf,
-        locks_dir: PathBuf,
-    ) -> Self {
-        Self {
-            api_client,
-            downloader: ParallelDownloader::new(blob_cache),
-            store,
-            cellar,
-            linker,
-            db,
-            prefix,
-            locks_dir,
-        }
-    }
-
-    /// Clear the API cache, forcing fresh metadata on next operation.
-    pub fn clear_api_cache(&self) -> Result<usize, Error> {
-        self.api_client.clear_cache()
-    }
-
-    /// Check if a specific installed package is outdated.
-    /// Returns Ok(None) if up-to-date, Ok(Some(..)) if outdated.
-    pub async fn is_outdated(&self, name: &str) -> Result<Option<OutdatedPackage>, Error> {
-        let installed = self.db.get_installed(name).ok_or(Error::NotInstalled {
-            name: name.to_string(),
-        })?;
-
-        let formula = self.api_client.get_formula(name).await?;
-        let is_source = installed.store_key.starts_with("source:");
-
-        if is_source {
-            let current_version = formula.effective_version();
-            if installed.version == current_version {
-                Ok(None)
-            } else {
-                Ok(Some(OutdatedPackage {
-                    name: name.to_string(),
-                    installed_version: installed.version,
-                    installed_sha256: installed.store_key,
-                    current_version,
-                    current_sha256: String::new(),
-                    is_source_build: true,
-                }))
-            }
-        } else {
-            let bottle = select_bottle(&formula)?;
-            if installed.store_key == bottle.sha256 {
-                Ok(None)
-            } else {
-                Ok(Some(OutdatedPackage {
-                    name: name.to_string(),
-                    installed_version: installed.version,
-                    installed_sha256: installed.store_key,
-                    current_version: formula.effective_version(),
-                    current_sha256: bottle.sha256,
-                    is_source_build: false,
-                }))
-            }
-        }
-    }
-
-    pub async fn check_outdated(&self) -> Result<(Vec<OutdatedPackage>, Vec<String>), Error> {
-        use std::collections::HashMap;
-
-        let installed = self.db.list_installed()?;
-        if installed.is_empty() {
-            return Ok((Vec::new(), Vec::new()));
-        }
-
-        let installed_names: std::collections::HashSet<&str> =
-            installed.iter().map(|k| k.name.as_str()).collect();
-
-        let bulk_raw = self.api_client.get_all_formulas_raw().await?;
-        let bulk_values: Vec<serde_json::Value> = serde_json::from_str(&bulk_raw)
-            .map_err(Error::network("failed to parse bulk formula JSON"))?;
-
-        let mut bulk_map: HashMap<String, Formula> = HashMap::new();
-        for val in bulk_values {
-            let name = match val.get("name").and_then(|n| n.as_str()) {
-                Some(n) if installed_names.contains(n) => n.to_string(),
-                _ => continue,
-            };
-            if let Ok(f) = serde_json::from_value(val) {
-                bulk_map.insert(name, f);
-            }
-        }
-
-        let mut outdated = Vec::new();
-        let mut warnings = Vec::new();
-
-        for keg in &installed {
-            let is_tap = keg.name.contains('/');
-
-            let formula = if is_tap || !bulk_map.contains_key(&keg.name) {
-                match self.api_client.get_formula(&keg.name).await {
-                    Ok(f) => f,
-                    Err(e) => {
-                        warnings.push(format!("{}: {}", keg.name, e));
-                        continue;
-                    }
-                }
-            } else {
-                bulk_map.remove(&keg.name).unwrap()
-            };
-
-            let is_source = keg.store_key.starts_with("source:");
-
-            if is_source {
-                let current_version = formula.effective_version();
-                if keg.version != current_version {
-                    outdated.push(OutdatedPackage {
-                        name: keg.name.clone(),
-                        installed_version: keg.version.clone(),
-                        installed_sha256: keg.store_key.clone(),
-                        current_version,
-                        current_sha256: String::new(),
-                        is_source_build: true,
-                    });
-                }
-            } else {
-                match select_bottle(&formula) {
-                    Ok(bottle) => {
-                        if keg.store_key != bottle.sha256 {
-                            outdated.push(OutdatedPackage {
-                                name: keg.name.clone(),
-                                installed_version: keg.version.clone(),
-                                installed_sha256: keg.store_key.clone(),
-                                current_version: formula.effective_version(),
-                                current_sha256: bottle.sha256,
-                                is_source_build: false,
-                            });
-                        }
-                    }
-                    Err(e) => warnings.push(format!("{}: {}", keg.name, e)),
-                }
-            }
-        }
-
-        outdated.sort_by(|a, b| a.name.cmp(&b.name));
-        Ok((outdated, warnings))
-    }
-
-    pub async fn suggest_formulas(&self, query: &str, limit: usize) -> Result<Vec<String>, Error> {
-        self.api_client.suggest_formulas(query, limit).await
-    }
-
-    pub async fn plan(&self, names: &[String]) -> Result<InstallPlan, Error> {
-        self.plan_with_options(names, false).await
-    }
-
-    pub async fn plan_with_options(
-        &self,
-        names: &[String],
-        build_from_source: bool,
-    ) -> Result<InstallPlan, Error> {
-        let formulas = self.fetch_all_formulas(names).await?;
-        let ordered = resolve_closure(names, &formulas)?;
-
-        let mut items = Vec::with_capacity(ordered.len());
-        for install_name in ordered {
-            let formula = formulas.get(&install_name).cloned().unwrap();
-            let method = if build_from_source {
-                match BuildPlan::from_formula(&formula, &self.prefix) {
-                    Some(plan) => InstallMethod::Source(plan),
-                    None => match select_bottle(&formula) {
-                        Ok(bottle) => InstallMethod::Bottle(bottle),
-                        Err(_) => {
-                            return Err(Error::UnsupportedBottle {
-                                name: formula.name.clone(),
-                            });
-                        }
-                    },
-                }
-            } else {
-                match select_bottle(&formula) {
-                    Ok(bottle) => InstallMethod::Bottle(bottle),
-                    Err(_) => match BuildPlan::from_formula(&formula, &self.prefix) {
-                        Some(plan) => InstallMethod::Source(plan),
-                        None => {
-                            return Err(Error::UnsupportedBottle {
-                                name: formula.name.clone(),
-                            });
-                        }
-                    },
-                }
-            };
-            items.push(PlannedInstall {
-                install_name,
-                formula,
-                method,
-            });
-        }
-
-        Ok(InstallPlan { items })
-    }
-
-    /// Try to extract a download, with automatic retry on corruption
-    async fn extract_with_retry(
-        &self,
-        download: &DownloadResult,
-        formula: &Formula,
-        bottle: &SelectedBottle,
-        progress: Option<DownloadProgressCallback>,
-    ) -> Result<std::path::PathBuf, Error> {
-        let mut blob_path = download.blob_path.clone();
-        let mut last_error = None;
-
-        for attempt in 0..MAX_CORRUPTION_RETRIES {
-            match self.store.ensure_entry(&bottle.sha256, &blob_path) {
-                Ok(entry) => return Ok(entry),
-                Err(Error::StoreCorruption { message }) => {
-                    // Remove the corrupted blob
-                    self.downloader.remove_blob(&bottle.sha256);
-
-                    if attempt + 1 < MAX_CORRUPTION_RETRIES {
-                        // Log retry attempt
-                        warn!(
-                            formula = %formula.name,
-                            attempt = attempt + 2,
-                            max_retries = MAX_CORRUPTION_RETRIES,
-                            "corrupted download detected; retrying"
-                        );
-
-                        // Re-download
-                        let request = DownloadRequest {
-                            url: bottle.url.clone(),
-                            sha256: bottle.sha256.clone(),
-                            name: formula.name.clone(),
-                        };
-
-                        match self
-                            .downloader
-                            .download_single(request, progress.clone())
-                            .await
-                        {
-                            Ok(new_path) => {
-                                blob_path = new_path;
-                                // Continue to next iteration to retry extraction
-                            }
-                            Err(e) => {
-                                last_error = Some(e);
-                                break;
-                            }
-                        }
-                    } else {
-                        last_error = Some(Error::StoreCorruption {
-                            message: format!(
-                                "{message}\n\nFailed after {MAX_CORRUPTION_RETRIES} attempts. The download may be corrupted at the source."
-                            ),
-                        });
-                    }
-                }
-                Err(e) => {
-                    last_error = Some(e);
-                    break;
-                }
-            }
-        }
-
-        Err(last_error.unwrap_or_else(|| Error::StoreCorruption {
-            message: "extraction failed with unknown error".to_string(),
-        }))
-    }
-
-    /// Recursively fetch a formula and all its dependencies in parallel batches
-    async fn fetch_all_formulas(
-        &self,
-        names: &[String],
-    ) -> Result<BTreeMap<String, Formula>, Error> {
-        use std::collections::HashSet;
-        use zb_core::select_bottle;
-
-        let mut formulas = BTreeMap::new();
-        let mut fetched: HashSet<String> = HashSet::new();
-        let mut to_fetch: Vec<String> = names.to_vec();
-
-        while !to_fetch.is_empty() {
-            // Fetch current batch in parallel
-            let batch: Vec<String> = to_fetch
-                .drain(..)
-                .filter(|n| !fetched.contains(n))
-                .collect();
-
-            if batch.is_empty() {
-                break;
-            }
-
-            // Mark as fetched before starting (to avoid re-queueing)
-            for n in &batch {
-                fetched.insert(n.clone());
-            }
-
-            // Fetch all in parallel
-            let futures: Vec<_> = batch
-                .iter()
-                .map(|n| self.api_client.get_formula(n))
-                .collect();
-
-            let results = futures::future::join_all(futures).await;
-
-            // Process results and queue new dependencies
-            for (i, result) in results.into_iter().enumerate() {
-                let formula = match result {
-                    Ok(f) => f,
-                    Err(e) => return Err(e),
-                };
-
-                if select_bottle(&formula).is_err() && !formula.has_source_url() {
-                    warn!(
-                        formula = %formula.name,
-                        "skipping formula with no bottle or source available for this platform"
-                    );
-                    continue;
-                }
-
-                // Queue dependencies for next batch
-                for dep in &formula.dependencies {
-                    if !fetched.contains(dep) && !to_fetch.contains(dep) {
-                        to_fetch.push(dep.clone());
-                    }
-                }
-
-                formulas.insert(batch[i].clone(), formula);
-            }
-        }
-
-        Ok(formulas)
-    }
-
-    /// Execute the install plan
-    pub async fn execute(&mut self, plan: InstallPlan, link: bool) -> Result<ExecuteResult, Error> {
-        self.execute_with_progress(plan, link, None).await
-    }
-
-    pub async fn execute_with_progress(
-        &mut self,
-        plan: InstallPlan,
-        link: bool,
-        progress: Option<Arc<ProgressCallback>>,
-    ) -> Result<ExecuteResult, Error> {
-        let lock_path = self.locks_dir.join("install.lock");
-        let lock_file =
-            File::create(&lock_path).map_err(Error::store("failed to create install lock"))?;
-        lock_file
-            .lock_exclusive()
-            .map_err(Error::store("failed to acquire install lock"))?;
-        let _lock = lock_file;
-
-        let report = |event: InstallProgress| {
-            if let Some(ref cb) = progress {
-                cb(event);
-            }
-        };
-
-        let (bottle_items, source_items): (Vec<_>, Vec<_>) = plan
-            .items
-            .into_iter()
-            .partition(|item| matches!(item.method, InstallMethod::Bottle(_)));
-
-        if bottle_items.is_empty() && source_items.is_empty() {
-            return Ok(ExecuteResult { installed: 0 });
-        }
-
-        let mut installed = 0usize;
-        let mut error: Option<Error> = None;
-
-        if !bottle_items.is_empty() {
-            let requests: Vec<DownloadRequest> = bottle_items
-                .iter()
-                .map(|item| {
-                    let InstallMethod::Bottle(ref bottle) = item.method else {
-                        unreachable!()
-                    };
-                    DownloadRequest {
-                        url: bottle.url.clone(),
-                        sha256: bottle.sha256.clone(),
-                        name: item.formula.name.clone(),
-                    }
-                })
-                .collect();
-
-            let download_progress: Option<DownloadProgressCallback> = progress.clone().map(|cb| {
-                Arc::new(move |event: InstallProgress| {
-                    cb(event);
-                }) as DownloadProgressCallback
-            });
-
-            let mut rx = self
-                .downloader
-                .download_streaming(requests, download_progress.clone());
-
-            while let Some(result) = rx.recv().await {
-                match result {
-                    Ok(download) => {
-                        match self
-                            .process_bottle_item(
-                                &bottle_items[download.index],
-                                &download,
-                                &download_progress,
-                                link,
-                                &report,
-                            )
-                            .await
-                        {
-                            Ok(()) => installed += 1,
-                            Err(e) => error = Some(e),
-                        }
-                    }
-                    Err(e) => {
-                        error = Some(e);
-                    }
-                }
-            }
-        }
-
-        for item in &source_items {
-            let InstallMethod::Source(ref build_plan) = item.method else {
-                unreachable!()
-            };
-
-            report(InstallProgress::UnpackStarted {
-                name: item.formula.name.clone(),
-            });
-
-            match self
-                .install_from_source(item, build_plan, link, &report)
-                .await
-            {
-                Ok(()) => installed += 1,
-                Err(e) => {
-                    error = Some(e);
-                    continue;
-                }
-            }
-        }
-
-        if let Some(e) = error {
-            return Err(e);
-        }
-
-        Ok(ExecuteResult { installed })
-    }
-
-    async fn process_bottle_item(
-        &mut self,
-        item: &PlannedInstall,
-        download: &DownloadResult,
-        download_progress: &Option<DownloadProgressCallback>,
-        link: bool,
-        report: &impl Fn(InstallProgress),
-    ) -> Result<(), Error> {
-        let InstallMethod::Bottle(ref bottle) = item.method else {
-            unreachable!()
-        };
-        let install_name = &item.install_name;
-        let formula_name = &item.formula.name;
-        let version = item.formula.effective_version();
-        let store_key = &bottle.sha256;
-
-        report(InstallProgress::UnpackStarted {
-            name: formula_name.clone(),
-        });
-
-        let store_entry = self
-            .extract_with_retry(download, &item.formula, bottle, download_progress.clone())
-            .await?;
-
-        let keg_path = self
-            .cellar
-            .materialize(formula_name, &version, &store_entry)?;
-
-        report(InstallProgress::UnpackCompleted {
-            name: formula_name.clone(),
-        });
-
-        let tx = self.db.transaction().inspect_err(|_| {
-            Self::cleanup_materialized(&self.cellar, formula_name, &version);
-        })?;
-
-        tx.record_install(install_name, &version, store_key)
-            .inspect_err(|_| {
-                Self::cleanup_materialized(&self.cellar, formula_name, &version);
-            })?;
-
-        tx.commit().inspect_err(|_| {
-            Self::cleanup_materialized(&self.cellar, formula_name, &version);
-        })?;
-
-        if let Err(e) = self.linker.link_opt(&keg_path) {
-            warn!(formula = %install_name, error = %e, "failed to create opt link");
-        }
-
-        if link && !item.formula.is_keg_only() {
-            report(InstallProgress::LinkStarted {
-                name: formula_name.clone(),
-            });
-            match self.linker.link_keg(&keg_path) {
-                Ok(linked_files) => {
-                    report(InstallProgress::LinkCompleted {
-                        name: formula_name.clone(),
-                    });
-                    self.record_linked_files(install_name, &version, &linked_files);
-                }
-                Err(e) => {
-                    let _ = self.linker.unlink_keg(&keg_path);
-                    report(InstallProgress::InstallCompleted {
-                        name: formula_name.clone(),
-                    });
-                    return Err(e);
-                }
-            }
-        } else if link && item.formula.is_keg_only() {
-            let reason = match &item.formula.keg_only {
-                zb_core::KegOnly::Reason(s) => s.clone(),
-                _ if formula_name.contains('@') => "versioned formula".to_string(),
-                _ => "keg-only formula".to_string(),
-            };
-            report(InstallProgress::LinkSkipped {
-                name: formula_name.clone(),
-                reason,
-            });
-        }
-
-        report(InstallProgress::InstallCompleted {
-            name: formula_name.clone(),
-        });
-
-        Ok(())
-    }
-
-    fn record_linked_files(
-        &mut self,
-        name: &str,
-        version: &str,
-        linked_files: &[crate::cellar::link::LinkedFile],
-    ) {
-        if let Ok(tx) = self.db.transaction() {
-            let mut ok = true;
-            for linked in linked_files {
-                if tx
-                    .record_linked_file(
-                        name,
-                        version,
-                        &linked.link_path.to_string_lossy(),
-                        &linked.target_path.to_string_lossy(),
-                    )
-                    .is_err()
-                {
-                    ok = false;
-                    break;
-                }
-            }
-            if ok {
-                let _ = tx.commit();
-            }
-        }
-    }
-
-    fn cleanup_failed_install(
-        linker: &Linker,
-        cellar: &Cellar,
-        name: &str,
-        version: &str,
-        keg_path: &Path,
-        unlink: bool,
-    ) {
-        if unlink && let Err(e) = linker.unlink_keg(keg_path) {
-            warn!(
-                formula = %name,
-                version = %version,
-                error = %e,
-                "failed to clean up links after install error"
-            );
-        }
-
-        if let Err(e) = cellar.remove_keg(name, version) {
-            warn!(
-                formula = %name,
-                version = %version,
-                error = %e,
-                "failed to remove keg after install error"
-            );
-        }
-    }
-
-    async fn install_from_source(
-        &mut self,
-        item: &PlannedInstall,
-        build_plan: &BuildPlan,
-        link: bool,
-        report: &impl Fn(InstallProgress),
-    ) -> Result<(), Error> {
-        let install_name = &item.install_name;
-        let formula_name = &item.formula.name;
-        let version = item.formula.effective_version();
-
-        let ruby_source_path =
-            item.formula
-                .ruby_source_path
-                .as_deref()
-                .ok_or_else(|| Error::ExecutionError {
-                    message: format!("no ruby_source_path for formula '{formula_name}'"),
-                })?;
-
-        let cache_dir = self.prefix.join("tmp").join("rb_cache");
-        let formula_rb_checksum = item
-            .formula
-            .ruby_source_checksum
-            .as_ref()
-            .map(|checksum| checksum.sha256.as_str());
-
-        let formula_rb = self
-            .api_client
-            .fetch_formula_rb(ruby_source_path, &cache_dir, formula_rb_checksum)
-            .await?;
-
-        let mut installed_deps = std::collections::HashMap::new();
-        for dep_name in &build_plan.runtime_dependencies {
-            if let Some(keg) = self.db.get_installed(dep_name) {
-                installed_deps.insert(
-                    dep_name.clone(),
-                    crate::build::DepInfo {
-                        cellar_path: dependency_cellar_path(&self.cellar, &keg.name, &keg.version),
-                    },
-                );
-            }
-        }
-
-        let keg_path = self.cellar.keg_path(formula_name, &version);
-        let previous_keg_backup =
-            Self::backup_existing_source_keg(&keg_path, formula_name, &version)?;
-
-        let executor = crate::build::BuildExecutor::new(self.prefix.clone());
-        if let Err(build_err) = executor
-            .execute(build_plan, &formula_rb, &installed_deps)
-            .await
-        {
-            if let Some(backup_path) = previous_keg_backup.as_ref() {
-                Self::restore_source_keg_from_backup(
-                    &keg_path,
-                    backup_path,
-                    formula_name,
-                    &version,
-                )?;
-            }
-            return Err(build_err);
-        }
-
-        if let Some(backup_path) = previous_keg_backup.as_ref() {
-            Self::remove_source_keg_backup(backup_path, formula_name, &version)?;
-        }
-
-        report(InstallProgress::UnpackCompleted {
-            name: formula_name.clone(),
-        });
-
-        let store_key = format!("source:{formula_name}:{version}");
-
-        let tx = self.db.transaction().inspect_err(|_| {
-            Self::cleanup_materialized(&self.cellar, formula_name, &version);
-        })?;
-
-        if let Err(e) = tx.record_install(install_name, &version, &store_key) {
-            drop(tx);
-            Self::cleanup_materialized(&self.cellar, formula_name, &version);
-            return Err(e);
-        }
-
-        if let Err(e) = tx.commit() {
-            Self::cleanup_materialized(&self.cellar, formula_name, &version);
-            return Err(e);
-        }
-
-        if let Err(e) = self.linker.link_opt(&keg_path) {
-            warn!(formula = %install_name, error = %e, "failed to create opt link");
-        }
-
-        let should_link = link && !item.formula.is_keg_only();
-
-        if should_link {
-            report(InstallProgress::LinkStarted {
-                name: formula_name.clone(),
-            });
-            match self.linker.link_keg(&keg_path) {
-                Ok(files) => {
-                    report(InstallProgress::LinkCompleted {
-                        name: formula_name.clone(),
-                    });
-                    if !files.is_empty()
-                        && let Ok(tx) = self.db.transaction()
-                    {
-                        let mut ok = true;
-                        for linked in &files {
-                            if tx
-                                .record_linked_file(
-                                    install_name,
-                                    &version,
-                                    &linked.link_path.to_string_lossy(),
-                                    &linked.target_path.to_string_lossy(),
-                                )
-                                .is_err()
-                            {
-                                ok = false;
-                                break;
-                            }
-                        }
-                        if ok {
-                            let _ = tx.commit();
-                        }
-                    }
-                }
-                Err(e) => {
-                    let _ = self.linker.unlink_keg(&keg_path);
-                    report(InstallProgress::InstallCompleted {
-                        name: formula_name.clone(),
-                    });
-                    return Err(e);
-                }
-            }
-        } else if link && item.formula.is_keg_only() {
-            let reason = match &item.formula.keg_only {
-                zb_core::KegOnly::Reason(s) => s.clone(),
-                _ if item.formula.name.contains('@') => "versioned formula".to_string(),
-                _ => "keg-only formula".to_string(),
-            };
-            report(InstallProgress::LinkSkipped {
-                name: formula_name.clone(),
-                reason,
-            });
-        }
-
-        report(InstallProgress::InstallCompleted {
-            name: formula_name.clone(),
-        });
-        Ok(())
-    }
-
-    fn backup_existing_source_keg(
-        keg_path: &Path,
-        formula_name: &str,
-        version: &str,
-    ) -> Result<Option<PathBuf>, Error> {
-        if !keg_path.exists() {
-            return Ok(None);
-        }
-
-        let backup_path = Self::source_keg_backup_path(keg_path);
-        if backup_path.exists() {
-            fs::remove_dir_all(&backup_path).map_err(|e| Error::StoreCorruption {
-                message: format!(
-                    "failed to remove stale source-build backup for '{}@{}': {}",
-                    formula_name, version, e
-                ),
-            })?;
-        }
-
-        fs::rename(keg_path, &backup_path).map_err(|e| Error::StoreCorruption {
-            message: format!(
-                "failed to backup existing keg for '{}@{}': {}",
-                formula_name, version, e
-            ),
-        })?;
-
-        Ok(Some(backup_path))
-    }
-
-    fn restore_source_keg_from_backup(
-        keg_path: &Path,
-        backup_path: &Path,
-        formula_name: &str,
-        version: &str,
-    ) -> Result<(), Error> {
-        if keg_path.exists() {
-            fs::remove_dir_all(keg_path).map_err(|e| Error::StoreCorruption {
-                message: format!(
-                    "failed to remove failed source-build output for '{}@{}': {}",
-                    formula_name, version, e
-                ),
-            })?;
-        }
-
-        fs::rename(backup_path, keg_path).map_err(|e| Error::StoreCorruption {
-            message: format!(
-                "failed to restore previous keg for '{}@{}': {}",
-                formula_name, version, e
-            ),
-        })
-    }
-
-    fn remove_source_keg_backup(
-        backup_path: &Path,
-        formula_name: &str,
-        version: &str,
-    ) -> Result<(), Error> {
-        if !backup_path.exists() {
-            return Ok(());
-        }
-
-        fs::remove_dir_all(backup_path).map_err(|e| Error::StoreCorruption {
-            message: format!(
-                "failed to remove source-build backup for '{}@{}': {}",
-                formula_name, version, e
-            ),
-        })
-    }
-
-    fn source_keg_backup_path(keg_path: &Path) -> PathBuf {
-        let backup_suffix = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_nanos())
-            .unwrap_or(0);
-        let name = keg_path
-            .file_name()
-            .map(|s| s.to_string_lossy().into_owned())
-            .unwrap_or_else(|| "keg".to_string());
-
-        keg_path.with_file_name(format!("{name}.zb-backup-{backup_suffix}"))
-    }
-
-    /// Remove a materialized keg that was never registered in the database.
-    fn cleanup_materialized(cellar: &Cellar, name: &str, version: &str) {
-        if let Err(e) = cellar.remove_keg(name, version) {
-            warn!(
-                formula = %name,
-                version = %version,
-                error = %e,
-                "failed to remove keg after install error"
-            );
-        }
-    }
-
-    /// Convenience method to plan and execute in one call
-    pub async fn install(&mut self, names: &[String], link: bool) -> Result<ExecuteResult, Error> {
-        let (casks, formulas): (Vec<_>, Vec<_>) = names
-            .iter()
-            .cloned()
-            .partition(|name| name.starts_with("cask:"));
-
-        let mut installed = 0usize;
-
-        if !formulas.is_empty() {
-            let plan = self.plan(&formulas).await?;
-            installed += self.execute(plan, link).await?.installed;
-        }
-
-        if !casks.is_empty() {
-            installed += self.install_casks(&casks, link).await?.installed;
-        }
-
-        Ok(ExecuteResult { installed })
-    }
-
-    pub async fn install_casks(
-        &mut self,
-        names: &[String],
-        link: bool,
-    ) -> Result<ExecuteResult, Error> {
-        let mut installed = 0usize;
-        for name in names {
-            let token = name
-                .strip_prefix("cask:")
-                .expect("install_casks expects cask: prefixed names");
-            self.install_single_cask(token, link).await?;
-            installed += 1;
-        }
-        Ok(ExecuteResult { installed })
-    }
-
-    /// Uninstall a formula
-    pub fn uninstall(&mut self, name: &str) -> Result<(), Error> {
-        // Check if installed
-        let installed = self.db.get_installed(name).ok_or(Error::NotInstalled {
-            name: name.to_string(),
-        })?;
-        let keg_name = formula_token(&installed.name);
-
-        // Unlink executables
-        let keg_path = self.cellar.keg_path(keg_name, &installed.version);
-        self.linker.unlink_keg(&keg_path)?;
-
-        // Remove from database (decrements store ref)
-        {
-            let tx = self.db.transaction()?;
-            tx.record_uninstall(name)?;
-            tx.commit()?;
-        }
-
-        // Remove cellar entry
-        self.cellar.remove_keg(keg_name, &installed.version)?;
-
-        Ok(())
-    }
-
-    /// Garbage collect unreferenced store entries
-    pub fn gc(&mut self) -> Result<Vec<String>, Error> {
-        let unreferenced = self.db.get_unreferenced_store_keys()?;
-        let mut removed = Vec::new();
-
-        for store_key in unreferenced {
-            self.store.remove_entry(&store_key)?;
-            self.db.delete_store_ref(&store_key)?;
-            removed.push(store_key);
-        }
-
-        Ok(removed)
-    }
-
-    /// Check if a formula is installed
-    pub fn is_installed(&self, name: &str) -> bool {
-        self.db.get_installed(name).is_some()
-    }
-
-    /// Get info about an installed formula
-    pub fn get_installed(&self, name: &str) -> Option<crate::storage::db::InstalledKeg> {
-        self.db.get_installed(name)
-    }
-
-    /// List all installed formulas
-    pub fn list_installed(&self) -> Result<Vec<crate::storage::db::InstalledKeg>, Error> {
-        self.db.list_installed()
-    }
-
-    /// Get the path to a keg in the cellar
-    pub fn keg_path(&self, name: &str, version: &str) -> std::path::PathBuf {
-        self.cellar.keg_path(name, version)
-    }
-    async fn install_single_cask(&mut self, token: &str, link: bool) -> Result<(), Error> {
-        let cask_json = self.api_client.get_cask(token).await?;
-        let cask = resolve_cask(token, &cask_json)?;
-
-        let blob_path = self
-            .downloader
-            .download_single(
-                DownloadRequest {
-                    url: cask.url.clone(),
-                    sha256: cask.sha256.clone(),
-                    name: cask.install_name.clone(),
-                },
-                None,
-            )
-            .await?;
-
-        let keg_path = self.cellar.keg_path(&cask.install_name, &cask.version);
-        let mut cleanup = FailedInstallGuard::new(
-            &self.linker,
-            &self.cellar,
-            &cask.install_name,
-            &cask.version,
-            &keg_path,
-            link,
-        );
-
-        if crate::extraction::is_archive(&blob_path)? {
-            let extracted = self.store.ensure_entry(&cask.sha256, &blob_path)?;
-            stage_cask_binaries(&extracted, &keg_path, &cask)?;
-        } else {
-            stage_raw_cask_binary(&blob_path, &keg_path, &cask)?;
-        }
-
-        let linked_files = if link {
-            self.linker.link_keg(&keg_path)?
-        } else {
-            Vec::new()
-        };
-
-        let tx = self.db.transaction()?;
-        tx.record_install(&cask.install_name, &cask.version, &cask.sha256)?;
-        for linked in &linked_files {
-            tx.record_linked_file(
-                &cask.install_name,
-                &cask.version,
-                &linked.link_path.to_string_lossy(),
-                &linked.target_path.to_string_lossy(),
-            )?;
-        }
-        tx.commit()?;
-
-        cleanup.disarm();
-        Ok(())
-    }
-}
-
-fn dependency_cellar_path(cellar: &Cellar, installed_name: &str, version: &str) -> String {
-    cellar
-        .keg_path(formula_token(installed_name), version)
-        .display()
-        .to_string()
-}
-
-struct FailedInstallGuard<'a> {
-    linker: &'a Linker,
-    cellar: &'a Cellar,
-    name: &'a str,
-    version: &'a str,
-    keg_path: &'a Path,
-    unlink: bool,
-    armed: bool,
-}
-
-impl<'a> FailedInstallGuard<'a> {
-    fn new(
-        linker: &'a Linker,
-        cellar: &'a Cellar,
-        name: &'a str,
-        version: &'a str,
-        keg_path: &'a Path,
-        unlink: bool,
-    ) -> Self {
-        Self {
-            linker,
-            cellar,
-            name,
-            version,
-            keg_path,
-            unlink,
-            armed: true,
-        }
-    }
-
-    fn disarm(&mut self) {
-        self.armed = false;
-    }
-}
-
-impl Drop for FailedInstallGuard<'_> {
-    fn drop(&mut self) {
-        if self.armed {
-            Installer::cleanup_failed_install(
-                self.linker,
-                self.cellar,
-                self.name,
-                self.version,
-                self.keg_path,
-                self.unlink,
-            );
-        }
-    }
-}
-
-fn stage_cask_binaries(
-    extracted_root: &Path,
-    keg_path: &Path,
-    cask: &crate::installer::cask::ResolvedCask,
-) -> Result<(), Error> {
-    let bin_dir = keg_path.join("bin");
-    fs::create_dir_all(&bin_dir).map_err(Error::store("failed to create cask bin dir"))?;
-
-    for binary in &cask.binaries {
-        let source = resolve_cask_source_path(extracted_root, cask, &binary.source)?;
-        if !source.exists() {
-            return Err(Error::InvalidArgument {
-                message: format!(
-                    "cask '{}' binary source '{}' not found in archive",
-                    cask.token, binary.source
-                ),
-            });
-        }
-
-        let target = bin_dir.join(&binary.target);
-        if target.exists() {
-            fs::remove_file(&target)
-                .map_err(Error::store("failed to replace existing cask binary"))?;
-        }
-
-        fs::copy(&source, &target).map_err(|e| Error::StoreCorruption {
-            message: format!("failed to stage cask binary '{}': {e}", binary.target),
-        })?;
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let mut perms = fs::metadata(&target)
-                .map_err(Error::store("failed to read staged cask binary metadata"))?
-                .permissions();
-            if perms.mode() & 0o111 == 0 {
-                perms.set_mode(0o755);
-                fs::set_permissions(&target, perms)
-                    .map_err(Error::store("failed to make staged cask binary executable"))?;
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn stage_raw_cask_binary(
-    blob_path: &Path,
-    keg_path: &Path,
-    cask: &crate::installer::cask::ResolvedCask,
-) -> Result<(), Error> {
-    if cask.binaries.len() != 1 {
-        return Err(Error::InvalidArgument {
-            message: format!(
-                "cask '{}' has {} binary artifacts but the download is a raw binary; expected exactly 1",
-                cask.token,
-                cask.binaries.len()
-            ),
-        });
-    }
-
-    let binary = &cask.binaries[0];
-    let bin_dir = keg_path.join("bin");
-    fs::create_dir_all(&bin_dir).map_err(Error::store("failed to create cask bin dir"))?;
-
-    let target = bin_dir.join(&binary.target);
-    if target.exists() {
-        fs::remove_file(&target).map_err(Error::store("failed to replace existing cask binary"))?;
-    }
-
-    fs::copy(blob_path, &target).map_err(|e| Error::StoreCorruption {
-        message: format!("failed to stage cask binary '{}': {e}", binary.target),
-    })?;
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        fs::set_permissions(&target, fs::Permissions::from_mode(0o755))
-            .map_err(Error::store("failed to make staged cask binary executable"))?;
-    }
-
-    Ok(())
-}
-
-fn resolve_cask_source_path(
-    extracted_root: &Path,
-    cask: &crate::installer::cask::ResolvedCask,
-    source: &str,
-) -> Result<std::path::PathBuf, Error> {
-    if source.starts_with("$APPDIR") {
-        return Err(Error::InvalidArgument {
-            message: format!(
-                "cask '{}' uses APPDIR artifacts which are not supported yet",
-                cask.token
-            ),
-        });
-    }
-
-    let mut normalized = source.to_string();
-    let caskroom_prefix = format!("$HOMEBREW_PREFIX/Caskroom/{}/{}/", cask.token, cask.version);
-    if let Some(stripped) = normalized.strip_prefix(&caskroom_prefix) {
-        normalized = stripped.to_string();
-    }
-
-    let source_path = Path::new(&normalized);
-    if source_path.is_absolute() {
-        return Err(Error::InvalidArgument {
-            message: format!(
-                "cask '{}' binary source '{}' must be a relative path",
-                cask.token, source
-            ),
-        });
-    }
-
-    for component in source_path.components() {
-        if matches!(component, std::path::Component::ParentDir) {
-            return Err(Error::InvalidArgument {
-                message: format!(
-                    "cask '{}' binary source '{}' cannot contain '..'",
-                    cask.token, source
-                ),
-            });
-        }
-    }
-
-    Ok(extracted_root.join(source_path))
-}
-
-/// Create an Installer with standard paths
-pub fn create_installer(
-    root: &Path,
-    prefix: &Path,
-    concurrency: usize,
-) -> Result<Installer, Error> {
-    use std::fs;
-
-    // First ensure the root directory exists
-    if !root.exists() {
-        fs::create_dir_all(root).map_err(|e| {
-            if e.kind() == std::io::ErrorKind::PermissionDenied {
-                Error::StoreCorruption {
-                    message: format!(
-                        "cannot create root directory '{}': permission denied.\n\n\
-                        Create it with:\n  sudo mkdir -p {} && sudo chown $USER {}",
-                        root.display(),
-                        root.display(),
-                        root.display()
-                    ),
-                }
-            } else {
-                Error::StoreCorruption {
-                    message: format!("failed to create root directory '{}': {e}", root.display()),
-                }
-            }
-        })?;
-    }
-
-    // Ensure all subdirectories exist
-    fs::create_dir_all(root.join("db")).map_err(Error::store("failed to create db directory"))?;
-
-    fs::create_dir_all(root.join("cache"))
-        .map_err(Error::store("failed to create cache directory"))?;
-
-    let api_cache_path = root.join("cache/api-cache.sqlite");
-    let api_cache =
-        ApiCache::open(&api_cache_path).map_err(Error::store("failed to open API cache"))?;
-
-    let api_client = match std::env::var("ZEROBREW_API_URL") {
-        Ok(url) => ApiClient::with_base_url(url)?,
-        Err(_) => ApiClient::new(),
-    }
-    .with_cache(api_cache);
-
-    let blob_cache =
-        BlobCache::new(&root.join("cache")).map_err(Error::store("failed to create blob cache"))?;
-    let store = Store::new(root).map_err(Error::store("failed to create store"))?;
-    // Use prefix/Cellar so bottles' hardcoded rpaths work
-    let cellar =
-        Cellar::new_at(prefix.join("Cellar")).map_err(Error::store("failed to create cellar"))?;
-    let linker = Linker::new(prefix).map_err(Error::store("failed to create linker"))?;
-    let db = Database::open(&root.join("db/zb.sqlite3"))?;
-
-    let locks_dir = root.join("locks");
-    fs::create_dir_all(&locks_dir).map_err(Error::store("failed to create locks directory"))?;
-
-    use crate::network::download::ParallelDownloader;
-    let parallel_downloader = ParallelDownloader::with_concurrency(blob_cache, concurrency);
-
-    Ok(Installer {
-        api_client,
-        downloader: parallel_downloader,
-        store,
-        cellar,
-        linker,
-        db,
-        prefix: prefix.to_path_buf(),
-        locks_dir,
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-    use tempfile::TempDir;
-    use wiremock::matchers::{method, path};
-    use wiremock::{Mock, MockServer, ResponseTemplate};
-
-    fn create_bottle_tarball(formula_name: &str) -> Vec<u8> {
-        use flate2::Compression;
-        use flate2::write::GzEncoder;
-        use std::io::Write;
-        use tar::Builder;
-
-        let mut builder = Builder::new(Vec::new());
-
-        // Create bin directory with executable
-        let mut header = tar::Header::new_gnu();
-        header
-            .set_path(format!("{}/1.0.0/bin/{}", formula_name, formula_name))
-            .unwrap();
-        header.set_size(20);
-        header.set_mode(0o755);
-        header.set_cksum();
-
-        let content = format!("#!/bin/sh\necho {}", formula_name);
-        builder.append(&header, content.as_bytes()).unwrap();
-
-        let tar_data = builder.into_inner().unwrap();
-
-        let mut encoder = GzEncoder::new(Vec::new(), Compression::default());
-        encoder.write_all(&tar_data).unwrap();
-        encoder.finish().unwrap()
-    }
-
-    fn sha256_hex(data: &[u8]) -> String {
-        use sha2::{Digest, Sha256};
-        let mut hasher = Sha256::new();
-        hasher.update(data);
-        format!("{:x}", hasher.finalize())
-    }
-
-    fn get_test_bottle_tag() -> &'static str {
-        if cfg!(target_os = "linux") {
-            "x86_64_linux"
-        } else if cfg!(target_arch = "x86_64") {
-            "sonoma"
-        } else {
-            "arm64_sonoma"
-        }
-    }
-
-    #[test]
-    fn dependency_cellar_path_uses_formula_token_for_tap_name() {
-        let tmp = TempDir::new().unwrap();
-        let cellar = Cellar::new(tmp.path()).unwrap();
-        let path = dependency_cellar_path(&cellar, "hashicorp/tap/terraform", "1.10.0");
-
-        assert!(path.ends_with("cellar/terraform/1.10.0"));
-    }
-
-    #[test]
-    fn dependency_cellar_path_keeps_core_formula_name() {
-        let tmp = TempDir::new().unwrap();
-        let cellar = Cellar::new(tmp.path()).unwrap();
-        let path = dependency_cellar_path(&cellar, "openssl@3", "3.3.2");
-
-        assert!(path.ends_with("cellar/openssl@3/3.3.2"));
-    }
-
-    #[test]
-    fn dependency_cellar_path_uses_name_from_db_record() {
-        let tmp = TempDir::new().unwrap();
-        let cellar = Cellar::new(tmp.path()).unwrap();
-
-        let db_path = tmp.path().join("zb.sqlite3");
-        let mut db = Database::open(&db_path).unwrap();
-        let tx = db.transaction().unwrap();
-        tx.record_install("hashicorp/tap/terraform", "1.10.0", "store-key")
-            .unwrap();
-        tx.commit().unwrap();
-
-        let keg = db.get_installed("hashicorp/tap/terraform").unwrap();
-        let path = dependency_cellar_path(&cellar, &keg.name, &keg.version);
-
-        assert!(path.ends_with("cellar/terraform/1.10.0"));
-    }
-
-    #[test]
-    fn source_keg_backup_can_restore_previous_installation() {
-        let tmp = TempDir::new().unwrap();
-        let keg_path = tmp.path().join("cellar").join("example").join("1.0.0");
-        fs::create_dir_all(&keg_path).unwrap();
-        fs::write(keg_path.join("old.txt"), "old").unwrap();
-
-        let backup = Installer::backup_existing_source_keg(&keg_path, "example", "1.0.0").unwrap();
-        let backup = backup.expect("backup path should exist");
-
-        assert!(!keg_path.exists());
-        assert!(backup.exists());
-
-        fs::create_dir_all(&keg_path).unwrap();
-        fs::write(keg_path.join("new.txt"), "new").unwrap();
-
-        Installer::restore_source_keg_from_backup(&keg_path, &backup, "example", "1.0.0").unwrap();
-
-        assert!(keg_path.join("old.txt").exists());
-        assert!(!keg_path.join("new.txt").exists());
-        assert!(!backup.exists());
-    }
-
-    #[test]
-    fn backup_existing_source_keg_returns_none_when_keg_is_missing() {
-        let tmp = TempDir::new().unwrap();
-        let missing_keg = tmp.path().join("cellar").join("example").join("1.0.0");
-
-        let backup =
-            Installer::backup_existing_source_keg(&missing_keg, "example", "1.0.0").unwrap();
-
-        assert!(backup.is_none());
-    }
-
-    #[tokio::test]
-    async fn suggest_formulas_returns_matches_from_api_client() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let bulk = r#"[
-            {"name":"python"},
-            {"name":"pytest"},
-            {"name":"pypy"}
-        ]"#;
-
-        Mock::given(method("GET"))
-            .and(path("/formula.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(bulk))
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix,
-            root.join("locks"),
-        );
-
-        let suggestions = installer.suggest_formulas("pythn", 3).await.unwrap();
-        assert_eq!(suggestions.first().map(String::as_str), Some("python"));
-    }
-
-    #[tokio::test]
-    async fn install_completes_successfully() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        // Create bottle
-        let bottle = create_bottle_tarball("testpkg");
-        let bottle_sha = sha256_hex(&bottle);
-
-        // Create formula JSON
-        let tag = get_test_bottle_tag();
-        let formula_json = format!(
-            r#"{{
-                "name": "testpkg",
-                "versions": {{ "stable": "1.0.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/testpkg-1.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            bottle_sha
-        );
-
-        // Mount formula API mock
-        Mock::given(method("GET"))
-            .and(path("/formula/testpkg.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&formula_json))
-            .mount(&mock_server)
-            .await;
-
-        // Mount bottle download mock
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/bottles/testpkg-1.0.0.{}.bottle.tar.gz",
-                tag
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(bottle.clone()))
-            .mount(&mock_server)
-            .await;
-
-        // Create installer with mocked API
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Install
-        installer
-            .install(&["testpkg".to_string()], true)
-            .await
-            .unwrap();
-
-        // Verify keg exists
-        assert!(root.join("cellar/testpkg/1.0.0").exists());
-
-        // Verify link exists
-        assert!(prefix.join("bin/testpkg").exists());
-
-        // Verify database records
-        let installed = installer.db.get_installed("testpkg");
-        assert!(installed.is_some());
-        assert_eq!(installed.unwrap().version, "1.0.0");
-    }
-
-    #[tokio::test]
-    async fn uninstall_cleans_everything() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        // Create bottle
-        let bottle = create_bottle_tarball("uninstallme");
-        let bottle_sha = sha256_hex(&bottle);
-
-        // Create formula JSON
-        let tag = get_test_bottle_tag();
-        let formula_json = format!(
-            r#"{{
-                "name": "uninstallme",
-                "versions": {{ "stable": "1.0.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/uninstallme-1.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            bottle_sha
-        );
-
-        // Mount mocks
-        Mock::given(method("GET"))
-            .and(path("/formula/uninstallme.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&formula_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/bottles/uninstallme-1.0.0.{}.bottle.tar.gz",
-                tag
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(bottle.clone()))
-            .mount(&mock_server)
-            .await;
-
-        // Create installer
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Install
-        installer
-            .install(&["uninstallme".to_string()], true)
-            .await
-            .unwrap();
-
-        // Verify installed
-        assert!(installer.is_installed("uninstallme"));
-        assert!(root.join("cellar/uninstallme/1.0.0").exists());
-        assert!(prefix.join("bin/uninstallme").exists());
-
-        // Uninstall
-        installer.uninstall("uninstallme").unwrap();
-
-        // Verify everything cleaned up
-        assert!(!installer.is_installed("uninstallme"));
-        assert!(!root.join("cellar/uninstallme/1.0.0").exists());
-        assert!(!prefix.join("bin/uninstallme").exists());
-    }
-
-    #[tokio::test]
-    async fn gc_removes_unreferenced_store_entries() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        // Create bottle
-        let bottle = create_bottle_tarball("gctest");
-        let bottle_sha = sha256_hex(&bottle);
-
-        // Create formula JSON
-        let tag = get_test_bottle_tag();
-        let formula_json = format!(
-            r#"{{
-                "name": "gctest",
-                "versions": {{ "stable": "1.0.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/gctest-1.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            bottle_sha
-        );
-
-        // Mount mocks
-        Mock::given(method("GET"))
-            .and(path("/formula/gctest.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&formula_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!("/bottles/gctest-1.0.0.{}.bottle.tar.gz", tag)))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(bottle.clone()))
-            .mount(&mock_server)
-            .await;
-
-        // Create installer
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Install and uninstall
-        installer
-            .install(&["gctest".to_string()], true)
-            .await
-            .unwrap();
-
-        // Store entry should exist before GC
-        assert!(root.join("store").join(&bottle_sha).exists());
-
-        installer.uninstall("gctest").unwrap();
-
-        // Store entry should still exist (refcount decremented but not GC'd)
-        assert!(root.join("store").join(&bottle_sha).exists());
-
-        // Run GC
-        let removed = installer.gc().unwrap();
-        assert_eq!(removed.len(), 1);
-        assert_eq!(removed[0], bottle_sha);
-
-        // Store entry should now be gone
-        assert!(!root.join("store").join(&bottle_sha).exists());
-        assert!(
-            installer
-                .db
-                .get_unreferenced_store_keys()
-                .unwrap()
-                .is_empty()
-        );
-    }
-
-    #[tokio::test]
-    async fn gc_does_not_remove_referenced_store_entries() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        // Create bottle
-        let bottle = create_bottle_tarball("keepme");
-        let bottle_sha = sha256_hex(&bottle);
-
-        // Create formula JSON
-        let tag = get_test_bottle_tag();
-        let formula_json = format!(
-            r#"{{
-                "name": "keepme",
-                "versions": {{ "stable": "1.0.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/keepme-1.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            bottle_sha
-        );
-
-        // Mount mocks
-        Mock::given(method("GET"))
-            .and(path("/formula/keepme.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&formula_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!("/bottles/keepme-1.0.0.{}.bottle.tar.gz", tag)))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(bottle.clone()))
-            .mount(&mock_server)
-            .await;
-
-        // Create installer
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Install but don't uninstall
-        installer
-            .install(&["keepme".to_string()], true)
-            .await
-            .unwrap();
-
-        // Store entry should exist
-        assert!(root.join("store").join(&bottle_sha).exists());
-
-        // Run GC - should not remove anything
-        let removed = installer.gc().unwrap();
-        assert!(removed.is_empty());
-
-        // Store entry should still exist
-        assert!(root.join("store").join(&bottle_sha).exists());
-    }
-
-    #[tokio::test]
-    async fn install_with_dependencies() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        // Create bottles
-        let dep_bottle = create_bottle_tarball("deplib");
-        let dep_sha = sha256_hex(&dep_bottle);
-
-        let main_bottle = create_bottle_tarball("mainpkg");
-        let main_sha = sha256_hex(&main_bottle);
-
-        // Create formula JSONs
-        let tag = get_test_bottle_tag();
-        let dep_json = format!(
-            r#"{{
-                "name": "deplib",
-                "versions": {{ "stable": "1.0.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/deplib-1.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            dep_sha
-        );
-
-        let main_json = format!(
-            r#"{{
-                "name": "mainpkg",
-                "versions": {{ "stable": "2.0.0" }},
-                "dependencies": ["deplib"],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/mainpkg-2.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            main_sha
-        );
-
-        // Mount mocks
-        Mock::given(method("GET"))
-            .and(path("/formula/deplib.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&dep_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path("/formula/mainpkg.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&main_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!("/bottles/deplib-1.0.0.{}.bottle.tar.gz", tag)))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(dep_bottle))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/bottles/mainpkg-2.0.0.{}.bottle.tar.gz",
-                tag
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(main_bottle))
-            .mount(&mock_server)
-            .await;
-
-        // Create installer
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Install main package (should also install dependency)
-        installer
-            .install(&["mainpkg".to_string()], true)
-            .await
-            .unwrap();
-
-        // Both packages should be installed
-        assert!(installer.db.get_installed("mainpkg").is_some());
-        assert!(installer.db.get_installed("deplib").is_some());
-    }
-
-    #[tokio::test]
-    async fn plans_tapped_formula_with_core_dependency() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let dep_bottle = create_bottle_tarball("go");
-        let dep_sha = sha256_hex(&dep_bottle);
-        let tag = get_test_bottle_tag();
-        let dep_json = format!(
-            r#"{{
-                "name": "go",
-                "versions": {{ "stable": "1.24.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/go-1.24.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            dep_sha
-        );
-
-        Mock::given(method("GET"))
-            .and(path("/formula/go.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&dep_json))
-            .mount(&mock_server)
-            .await;
-
-        let tap_formula_rb = format!(
-            r#"
-class Terraform < Formula
-  version "1.10.0"
-  depends_on "go"
-  bottle do
-    root_url "{}/ghcr/hashicorp/tap"
-    sha256 {}: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-  end
-end
-"#,
-            mock_server.uri(),
-            tag
-        );
-
-        Mock::given(method("GET"))
-            .and(path("/hashicorp/homebrew-tap/main/Formula/terraform.rb"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(tap_formula_rb))
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client = ApiClient::with_base_url(format!("{}/formula", mock_server.uri()))
-            .unwrap()
-            .with_tap_raw_base_url(mock_server.uri());
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.to_path_buf(),
-            root.join("locks"),
-        );
-        let plan = installer
-            .plan(&["hashicorp/tap/terraform".to_string()])
-            .await
-            .unwrap();
-
-        let planned_names: Vec<String> = plan
-            .items
-            .iter()
-            .map(|item| item.formula.name.clone())
-            .collect();
-        assert!(planned_names.contains(&"terraform".to_string()));
-        assert!(planned_names.contains(&"go".to_string()));
-    }
-
-    #[tokio::test]
-    async fn uninstall_accepts_full_tap_reference_after_install() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let bottle = create_bottle_tarball("terraform");
-        let sha = sha256_hex(&bottle);
-        let tag = get_test_bottle_tag();
-
-        let tap_formula_rb = format!(
-            r#"
-class Terraform < Formula
-  version "1.10.0"
-  bottle do
-    root_url "{}/v2/hashicorp/tap"
-    sha256 {}: "{}"
-  end
-end
-"#,
-            mock_server.uri(),
-            tag,
-            sha
-        );
-
-        Mock::given(method("GET"))
-            .and(path("/hashicorp/homebrew-tap/main/Formula/terraform.rb"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(tap_formula_rb))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/v2/hashicorp/tap/terraform/blobs/sha256:{sha}"
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(bottle))
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client = ApiClient::with_base_url(format!("{}/formula", mock_server.uri()))
-            .unwrap()
-            .with_tap_raw_base_url(mock_server.uri());
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.to_path_buf(),
-            root.join("locks"),
-        );
-
-        installer
-            .install(&["hashicorp/tap/terraform".to_string()], true)
-            .await
-            .unwrap();
-
-        assert!(installer.is_installed("hashicorp/tap/terraform"));
-        assert!(!installer.is_installed("terraform"));
-        assert!(root.join("cellar/terraform/1.10.0").exists());
-        installer.uninstall("hashicorp/tap/terraform").unwrap();
-        assert!(!installer.is_installed("hashicorp/tap/terraform"));
-        assert!(!root.join("cellar/terraform/1.10.0").exists());
-    }
-
-    #[tokio::test]
-    async fn uninstalling_non_installed_tap_ref_does_not_remove_core_formula() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let bottle = create_bottle_tarball("terraform");
-        let sha = sha256_hex(&bottle);
-        let tag = get_test_bottle_tag();
-        let core_json = format!(
-            r#"{{
-                "name": "terraform",
-                "versions": {{ "stable": "1.10.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/terraform-1.10.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            sha
-        );
-
-        Mock::given(method("GET"))
-            .and(path("/formula/terraform.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(core_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/bottles/terraform-1.10.0.{}.bottle.tar.gz",
-                tag
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(bottle))
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.to_path_buf(),
-            root.join("locks"),
-        );
-        installer
-            .install(&["terraform".to_string()], true)
-            .await
-            .unwrap();
-        assert!(installer.is_installed("terraform"));
-
-        let err = installer.uninstall("hashicorp/tap/terraform").unwrap_err();
-        assert!(matches!(err, Error::NotInstalled { .. }));
-        assert!(installer.is_installed("terraform"));
-    }
-
-    #[tokio::test]
-    async fn preserves_successful_installs_when_one_package_fails() {
-        use std::time::Duration;
-
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let good_bottle = create_bottle_tarball("goodpkg");
-        let good_sha = sha256_hex(&good_bottle);
-
-        let tag = get_test_bottle_tag();
-        let good_json = format!(
-            r#"{{
-                "name": "goodpkg",
-                "versions": {{ "stable": "1.0.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/goodpkg-1.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            good_sha
-        );
-
-        let bad_json = format!(
-            r#"{{
-                "name": "badpkg",
-                "versions": {{ "stable": "1.0.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/badpkg-1.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-        );
-
-        Mock::given(method("GET"))
-            .and(path("/formula/goodpkg.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&good_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path("/formula/badpkg.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&bad_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/bottles/goodpkg-1.0.0.{}.bottle.tar.gz",
-                tag
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(good_bottle))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!("/bottles/badpkg-1.0.0.{}.bottle.tar.gz", tag)))
-            .respond_with(
-                ResponseTemplate::new(500)
-                    .set_delay(Duration::from_millis(100))
-                    .set_body_string("download failed"),
-            )
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        let result = installer
-            .install(&["goodpkg".to_string(), "badpkg".to_string()], false)
-            .await;
-        assert!(result.is_err());
-
-        assert!(installer.db.get_installed("goodpkg").is_some());
-        assert!(installer.db.get_installed("badpkg").is_none());
-        assert!(root.join("cellar/goodpkg/1.0.0").exists());
-    }
-
-    #[tokio::test]
-    async fn db_persist_failure_cleans_materialized_and_linked_files() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let bottle = create_bottle_tarball("rollbackme");
-        let bottle_sha = sha256_hex(&bottle);
-
-        let tag = get_test_bottle_tag();
-        let formula_json = format!(
-            r#"{{
-                "name": "rollbackme",
-                "versions": {{ "stable": "1.0.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/rollbackme-1.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            bottle_sha
-        );
-
-        Mock::given(method("GET"))
-            .and(path("/formula/rollbackme.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&formula_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/bottles/rollbackme-1.0.0.{}.bottle.tar.gz",
-                tag
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(bottle))
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let db_path = root.join("db/zb.sqlite3");
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&db_path).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Force metadata persistence to fail after filesystem work is done.
-        let conn = rusqlite::Connection::open(&db_path).unwrap();
-        conn.execute("DROP TABLE installed_kegs", []).unwrap();
-
-        let result = installer.install(&["rollbackme".to_string()], true).await;
-        assert!(result.is_err());
-
-        assert!(!root.join("cellar/rollbackme/1.0.0").exists());
-        assert!(!prefix.join("bin/rollbackme").exists());
-        assert!(!prefix.join("opt/rollbackme").exists());
-        assert!(root.join("store").join(&bottle_sha).exists());
-    }
-
-    #[tokio::test]
-    async fn db_persist_failure_cleans_materialized_tap_formula_keg() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let bottle = create_bottle_tarball("terraform");
-        let bottle_sha = sha256_hex(&bottle);
-        let tag = get_test_bottle_tag();
-
-        let tap_formula_rb = format!(
-            r#"
-class Terraform < Formula
-  version "1.10.0"
-  bottle do
-    root_url "{}/v2/hashicorp/tap"
-    sha256 {}: "{}"
-  end
-end
-"#,
-            mock_server.uri(),
-            tag,
-            bottle_sha
-        );
-
-        Mock::given(method("GET"))
-            .and(path("/hashicorp/homebrew-tap/main/Formula/terraform.rb"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(tap_formula_rb))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/v2/hashicorp/tap/terraform/blobs/sha256:{bottle_sha}"
-            )))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(bottle))
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let db_path = root.join("db/zb.sqlite3");
-        let api_client = ApiClient::with_base_url(format!("{}/formula", mock_server.uri()))
-            .unwrap()
-            .with_tap_raw_base_url(mock_server.uri());
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&db_path).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Force metadata persistence to fail after filesystem work is done.
-        let conn = rusqlite::Connection::open(&db_path).unwrap();
-        conn.execute("DROP TABLE installed_kegs", []).unwrap();
-
-        let result = installer
-            .install(&["hashicorp/tap/terraform".to_string()], true)
-            .await;
-        assert!(result.is_err());
-
-        // Keg is materialized at canonical formula name, so rollback must remove this path.
-        assert!(!root.join("cellar/terraform/1.10.0").exists());
-        assert!(!prefix.join("bin/terraform").exists());
-        assert!(!prefix.join("opt/terraform").exists());
-        assert!(root.join("store").join(&bottle_sha).exists());
-    }
-
-    #[tokio::test]
-    async fn parallel_api_fetching_with_deep_deps() {
-        // Tests that parallel API fetching works with a deeper dependency tree:
-        // root -> mid1 -> leaf1
-        //      -> mid2 -> leaf2
-        //              -> leaf1 (shared)
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        // Create bottles
-        let leaf1_bottle = create_bottle_tarball("leaf1");
-        let leaf1_sha = sha256_hex(&leaf1_bottle);
-        let leaf2_bottle = create_bottle_tarball("leaf2");
-        let leaf2_sha = sha256_hex(&leaf2_bottle);
-        let mid1_bottle = create_bottle_tarball("mid1");
-        let mid1_sha = sha256_hex(&mid1_bottle);
-        let mid2_bottle = create_bottle_tarball("mid2");
-        let mid2_sha = sha256_hex(&mid2_bottle);
-        let root_bottle = create_bottle_tarball("root");
-        let root_sha = sha256_hex(&root_bottle);
-
-        // Formula JSONs
-        let tag = get_test_bottle_tag();
-        let leaf1_json = format!(
-            r#"{{"name":"leaf1","versions":{{"stable":"1.0.0"}},"dependencies":[],"bottle":{{"stable":{{"files":{{"{}":{{"url":"{}/bottles/leaf1.tar.gz","sha256":"{}"}}}}}}}}}}"#,
-            tag,
-            mock_server.uri(),
-            leaf1_sha
-        );
-        let leaf2_json = format!(
-            r#"{{"name":"leaf2","versions":{{"stable":"1.0.0"}},"dependencies":[],"bottle":{{"stable":{{"files":{{"{}":{{"url":"{}/bottles/leaf2.tar.gz","sha256":"{}"}}}}}}}}}}"#,
-            tag,
-            mock_server.uri(),
-            leaf2_sha
-        );
-        let mid1_json = format!(
-            r#"{{"name":"mid1","versions":{{"stable":"1.0.0"}},"dependencies":["leaf1"],"bottle":{{"stable":{{"files":{{"{}":{{"url":"{}/bottles/mid1.tar.gz","sha256":"{}"}}}}}}}}}}"#,
-            tag,
-            mock_server.uri(),
-            mid1_sha
-        );
-        let mid2_json = format!(
-            r#"{{"name":"mid2","versions":{{"stable":"1.0.0"}},"dependencies":["leaf1","leaf2"],"bottle":{{"stable":{{"files":{{"{}":{{"url":"{}/bottles/mid2.tar.gz","sha256":"{}"}}}}}}}}}}"#,
-            tag,
-            mock_server.uri(),
-            mid2_sha
-        );
-        let root_json = format!(
-            r#"{{"name":"root","versions":{{"stable":"1.0.0"}},"dependencies":["mid1","mid2"],"bottle":{{"stable":{{"files":{{"{}":{{"url":"{}/bottles/root.tar.gz","sha256":"{}"}}}}}}}}}}"#,
-            tag,
-            mock_server.uri(),
-            root_sha
-        );
-
-        // Mount all mocks
-        for (name, json) in [
-            ("leaf1", &leaf1_json),
-            ("leaf2", &leaf2_json),
-            ("mid1", &mid1_json),
-            ("mid2", &mid2_json),
-            ("root", &root_json),
-        ] {
-            Mock::given(method("GET"))
-                .and(path(format!("/formula/{}.json", name)))
-                .respond_with(ResponseTemplate::new(200).set_body_string(json))
-                .mount(&mock_server)
-                .await;
-        }
-        for (name, bottle) in [
-            ("leaf1", &leaf1_bottle),
-            ("leaf2", &leaf2_bottle),
-            ("mid1", &mid1_bottle),
-            ("mid2", &mid2_bottle),
-            ("root", &root_bottle),
-        ] {
-            Mock::given(method("GET"))
-                .and(path(format!("/bottles/{}.tar.gz", name)))
-                .respond_with(ResponseTemplate::new(200).set_body_bytes(bottle.clone()))
-                .mount(&mock_server)
-                .await;
-        }
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Install root (should install all 5 packages)
-        installer
-            .install(&["root".to_string()], true)
-            .await
-            .unwrap();
-
-        // All packages should be installed
-        assert!(installer.db.get_installed("root").is_some());
-        assert!(installer.db.get_installed("mid1").is_some());
-        assert!(installer.db.get_installed("mid2").is_some());
-        assert!(installer.db.get_installed("leaf1").is_some());
-        assert!(installer.db.get_installed("leaf2").is_some());
-    }
-
-    #[tokio::test]
-    async fn streaming_extraction_processes_as_downloads_complete() {
-        // Tests that streaming extraction works correctly by verifying
-        // packages with delayed downloads still get installed properly
-        use std::time::Duration;
-
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        // Create bottles
-        let fast_bottle = create_bottle_tarball("fastpkg");
-        let fast_sha = sha256_hex(&fast_bottle);
-        let slow_bottle = create_bottle_tarball("slowpkg");
-        let slow_sha = sha256_hex(&slow_bottle);
-
-        // Fast package formula
-        let tag = get_test_bottle_tag();
-        let fast_json = format!(
-            r#"{{"name":"fastpkg","versions":{{"stable":"1.0.0"}},"dependencies":[],"bottle":{{"stable":{{"files":{{"{}":{{"url":"{}/bottles/fast.tar.gz","sha256":"{}"}}}}}}}}}}"#,
-            tag,
-            mock_server.uri(),
-            fast_sha
-        );
-
-        // Slow package formula (depends on fast)
-        let slow_json = format!(
-            r#"{{"name":"slowpkg","versions":{{"stable":"1.0.0"}},"dependencies":["fastpkg"],"bottle":{{"stable":{{"files":{{"{}":{{"url":"{}/bottles/slow.tar.gz","sha256":"{}"}}}}}}}}}}"#,
-            tag,
-            mock_server.uri(),
-            slow_sha
-        );
-
-        // Mount API mocks
-        Mock::given(method("GET"))
-            .and(path("/formula/fastpkg.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&fast_json))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path("/formula/slowpkg.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&slow_json))
-            .mount(&mock_server)
-            .await;
-
-        // Fast bottle responds immediately
-        Mock::given(method("GET"))
-            .and(path("/bottles/fast.tar.gz"))
-            .respond_with(ResponseTemplate::new(200).set_body_bytes(fast_bottle.clone()))
-            .mount(&mock_server)
-            .await;
-
-        // Slow bottle has a delay (simulates slow network)
-        Mock::given(method("GET"))
-            .and(path("/bottles/slow.tar.gz"))
-            .respond_with(
-                ResponseTemplate::new(200)
-                    .set_body_bytes(slow_bottle.clone())
-                    .set_delay(Duration::from_millis(100)),
-            )
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Install slow package (which depends on fast)
-        // With streaming, fast should be extracted while slow is still downloading
-        installer
-            .install(&["slowpkg".to_string()], true)
-            .await
-            .unwrap();
-
-        // Both packages should be installed
-        assert!(installer.db.get_installed("fastpkg").is_some());
-        assert!(installer.db.get_installed("slowpkg").is_some());
-
-        // Verify kegs exist
-        assert!(root.join("cellar/fastpkg/1.0.0").exists());
-        assert!(root.join("cellar/slowpkg/1.0.0").exists());
-
-        // Verify links exist
-        assert!(prefix.join("bin/fastpkg").exists());
-        assert!(prefix.join("bin/slowpkg").exists());
-    }
-
-    #[tokio::test]
-    async fn retries_on_corrupted_download() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        // Create valid bottle
-        let bottle = create_bottle_tarball("retrypkg");
-        let bottle_sha = sha256_hex(&bottle);
-
-        // Create formula JSON
-        let tag = get_test_bottle_tag();
-        let formula_json = format!(
-            r#"{{
-                "name": "retrypkg",
-                "versions": {{ "stable": "1.0.0" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "{}/bottles/retrypkg-1.0.0.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag,
-            mock_server.uri(),
-            tag,
-            bottle_sha
-        );
-
-        // Mount formula API mock
-        Mock::given(method("GET"))
-            .and(path("/formula/retrypkg.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&formula_json))
-            .mount(&mock_server)
-            .await;
-
-        // Track download attempts
-        let attempt_count = Arc::new(AtomicUsize::new(0));
-        let attempt_clone = attempt_count.clone();
-        let valid_bottle = bottle.clone();
-
-        // First request returns corrupted data (wrong content but matches sha for download)
-        // This simulates CDN corruption where sha passes but tar is invalid
-        Mock::given(method("GET"))
-            .and(path(format!(
-                "/bottles/retrypkg-1.0.0.{}.bottle.tar.gz",
-                tag
-            )))
-            .respond_with(move |_: &wiremock::Request| {
-                let attempt = attempt_clone.fetch_add(1, Ordering::SeqCst);
-                if attempt == 0 {
-                    // First attempt: return corrupted data
-                    // We need to return data that has the right sha256 but is corrupt
-                    // Since we can't fake sha256, we'll return invalid tar that will fail extraction
-                    // But actually the sha256 check happens during download...
-                    // So we need to return the valid bottle (sha passes) but corrupt the blob after
-                    // This is tricky to test since corruption happens at tar level
-                    // For now, just return valid data - the retry mechanism will work in real scenarios
-                    ResponseTemplate::new(200).set_body_bytes(valid_bottle.clone())
-                } else {
-                    // Subsequent attempts: return valid bottle
-                    ResponseTemplate::new(200).set_body_bytes(valid_bottle.clone())
-                }
-            })
-            .mount(&mock_server)
-            .await;
-
-        // Create installer
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let mut installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        // Install - should succeed (first download is valid in this test)
-        installer
-            .install(&["retrypkg".to_string()], true)
-            .await
-            .unwrap();
-
-        // Verify installation succeeded
-        assert!(installer.is_installed("retrypkg"));
-        assert!(root.join("cellar/retrypkg/1.0.0").exists());
-        assert!(prefix.join("bin/retrypkg").exists());
-    }
-
-    #[tokio::test]
-    async fn fails_after_max_retries() {
-        // This test verifies that after MAX_CORRUPTION_RETRIES failed attempts,
-        // the installer gives up with an appropriate error message.
-        // Note: This is hard to test without mocking the store layer since
-        // corruption is detected during tar extraction, not during download.
-        // The retry mechanism is validated by the code structure.
-
-        // For a proper integration test, we would need to inject corruption
-        // into the blob cache after download but before extraction.
-        // This is left as a documentation of the expected behavior:
-        // - First attempt: download succeeds, extraction fails (corruption)
-        // - Second attempt: re-download, extraction fails (corruption)
-        // - Third attempt: re-download, extraction fails (corruption)
-        // - Returns error: "Failed after 3 attempts..."
-    }
-
-    #[tokio::test]
-    async fn plan_falls_back_to_source_when_no_bottle() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let formula_json = r#"{
-            "name": "nobottle",
-            "versions": { "stable": "1.0.0" },
-            "dependencies": [],
-            "build_dependencies": ["pkgconf"],
-            "urls": {
-                "stable": {
-                    "url": "https://example.com/nobottle-1.0.0.tar.gz",
-                    "checksum": "abc123"
-                }
-            },
-            "ruby_source_path": "Formula/n/nobottle.rb",
-            "bottle": { "stable": { "files": {} } }
-        }"#;
-
-        Mock::given(method("GET"))
-            .and(path("/formula/nobottle.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(formula_json))
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        let plan = installer.plan(&["nobottle".to_string()]).await.unwrap();
-
-        assert_eq!(plan.items.len(), 1);
-        assert_eq!(plan.items[0].formula.name, "nobottle");
-        assert!(matches!(
-            plan.items[0].method,
-            zb_core::InstallMethod::Source(_)
-        ));
-
-        if let zb_core::InstallMethod::Source(ref bp) = plan.items[0].method {
-            assert_eq!(bp.source_url, "https://example.com/nobottle-1.0.0.tar.gz");
-            assert_eq!(bp.formula_name, "nobottle");
-            assert_eq!(bp.build_dependencies, vec!["pkgconf"]);
-        }
-    }
-
-    #[tokio::test]
-    async fn plan_prefers_bottle_over_source() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let tag = get_test_bottle_tag();
-        let formula_json = format!(
-            r#"{{
-                "name": "hasboth",
-                "versions": {{ "stable": "2.0.0" }},
-                "dependencies": [],
-                "urls": {{
-                    "stable": {{
-                        "url": "https://example.com/hasboth-2.0.0.tar.gz",
-                        "checksum": "def456"
-                    }}
-                }},
-                "ruby_source_path": "Formula/h/hasboth.rb",
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "https://example.com/hasboth.bottle.tar.gz",
-                                "sha256": "aabbccdd"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            tag
-        );
-
-        Mock::given(method("GET"))
-            .and(path("/formula/hasboth.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(&formula_json))
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        let plan = installer.plan(&["hasboth".to_string()]).await.unwrap();
-
-        assert_eq!(plan.items.len(), 1);
-        assert!(matches!(
-            plan.items[0].method,
-            zb_core::InstallMethod::Bottle(_)
-        ));
-    }
-
-    #[tokio::test]
-    async fn plan_errors_when_no_bottle_and_no_source() {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-
-        let formula_json = r#"{
-            "name": "nothing",
-            "versions": { "stable": "1.0.0" },
-            "dependencies": [],
-            "bottle": { "stable": { "files": {} } }
-        }"#;
-
-        Mock::given(method("GET"))
-            .and(path("/formula/nothing.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(formula_json))
-            .mount(&mock_server)
-            .await;
-
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix.clone(),
-            root.join("locks"),
-        );
-
-        let result = installer.plan(&["nothing".to_string()]).await;
-        assert!(result.is_err());
-        assert!(matches!(
-            result.unwrap_err(),
-            zb_core::Error::MissingFormula { .. }
-        ));
-    }
-
-    /// Helper: create a minimal Installer backed by a mock server and temp dir.
-    /// Returns (installer, mock_server, tmp_dir).
-    async fn test_installer() -> (Installer, MockServer, TempDir) {
-        let mock_server = MockServer::start().await;
-        let tmp = TempDir::new().unwrap();
-        let root = tmp.path().join("zerobrew");
-        let prefix = tmp.path().join("homebrew");
-        fs::create_dir_all(root.join("db")).unwrap();
-
-        let api_client =
-            ApiClient::with_base_url(format!("{}/formula", mock_server.uri())).unwrap();
-        let blob_cache = BlobCache::new(&root.join("cache")).unwrap();
-        let store = Store::new(&root).unwrap();
-        let cellar = Cellar::new(&root).unwrap();
-        let linker = Linker::new(&prefix).unwrap();
-        let db = Database::open(&root.join("db/zb.sqlite3")).unwrap();
-
-        let installer = Installer::new(
-            api_client,
-            blob_cache,
-            store,
-            cellar,
-            linker,
-            db,
-            prefix,
-            root.join("locks"),
-        );
-        (installer, mock_server, tmp)
-    }
-
-    fn formula_json(name: &str, version: &str, sha256: &str) -> String {
-        let tag = get_test_bottle_tag();
-        format!(
-            r#"{{
-                "name": "{}",
-                "versions": {{ "stable": "{}" }},
-                "dependencies": [],
-                "bottle": {{
-                    "stable": {{
-                        "files": {{
-                            "{}": {{
-                                "url": "https://example.com/{}-{}.{}.bottle.tar.gz",
-                                "sha256": "{}"
-                            }}
-                        }}
-                    }}
-                }}
-            }}"#,
-            name, version, tag, name, version, tag, sha256
-        )
-    }
-
-    #[tokio::test]
-    async fn is_outdated_returns_none_when_sha256_matches() {
-        let (mut installer, mock_server, _tmp) = test_installer().await;
-        let sha = "abc123def456";
-
-        {
-            let tx = installer.db.transaction().unwrap();
-            tx.record_install("jq", "1.7.1", sha).unwrap();
-            tx.commit().unwrap();
-        }
-
-        Mock::given(method("GET"))
-            .and(path("/formula/jq.json"))
-            .respond_with(
-                ResponseTemplate::new(200).set_body_string(formula_json("jq", "1.7.1", sha)),
-            )
-            .mount(&mock_server)
-            .await;
-
-        let result = installer.is_outdated("jq").await.unwrap();
-        assert!(result.is_none());
-    }
-
-    #[tokio::test]
-    async fn is_outdated_returns_some_when_sha256_differs() {
-        let (mut installer, mock_server, _tmp) = test_installer().await;
-
-        {
-            let tx = installer.db.transaction().unwrap();
-            tx.record_install("jq", "1.7.0", "old_sha256").unwrap();
-            tx.commit().unwrap();
-        }
-
-        Mock::given(method("GET"))
-            .and(path("/formula/jq.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(formula_json(
-                "jq",
-                "1.7.1",
-                "new_sha256",
-            )))
-            .mount(&mock_server)
-            .await;
-
-        let result = installer.is_outdated("jq").await.unwrap().unwrap();
-        assert_eq!(result.name, "jq");
-        assert_eq!(result.installed_version, "1.7.0");
-        assert_eq!(result.current_version, "1.7.1");
-        assert!(!result.is_source_build);
-    }
-
-    #[tokio::test]
-    async fn is_outdated_errors_for_not_installed() {
-        let (installer, _mock_server, _tmp) = test_installer().await;
-
-        let err = installer.is_outdated("jq").await.unwrap_err();
-        assert!(matches!(err, zb_core::Error::NotInstalled { .. }));
-    }
-
-    #[tokio::test]
-    async fn is_outdated_source_build_compares_version_only() {
-        let (mut installer, mock_server, _tmp) = test_installer().await;
-
-        {
-            let tx = installer.db.transaction().unwrap();
-            tx.record_install("jq", "1.7.1", "source:jq:1.7.1").unwrap();
-            tx.commit().unwrap();
-        }
-
-        Mock::given(method("GET"))
-            .and(path("/formula/jq.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(formula_json(
-                "jq",
-                "1.7.1",
-                "irrelevant",
-            )))
-            .mount(&mock_server)
-            .await;
-
-        let result = installer.is_outdated("jq").await.unwrap();
-        assert!(result.is_none());
-    }
-
-    #[tokio::test]
-    async fn is_outdated_source_build_detects_new_version() {
-        let (mut installer, mock_server, _tmp) = test_installer().await;
-
-        {
-            let tx = installer.db.transaction().unwrap();
-            tx.record_install("jq", "1.6", "source:jq:1.6").unwrap();
-            tx.commit().unwrap();
-        }
-
-        Mock::given(method("GET"))
-            .and(path("/formula/jq.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(formula_json(
-                "jq",
-                "1.7.1",
-                "irrelevant",
-            )))
-            .mount(&mock_server)
-            .await;
-
-        let result = installer.is_outdated("jq").await.unwrap().unwrap();
-        assert_eq!(result.installed_version, "1.6");
-        assert_eq!(result.current_version, "1.7.1");
-        assert!(result.is_source_build);
-    }
-
-    #[tokio::test]
-    async fn check_outdated_empty_when_nothing_installed() {
-        let (installer, _mock_server, _tmp) = test_installer().await;
-
-        let (outdated, warnings) = installer.check_outdated().await.unwrap();
-        assert!(outdated.is_empty());
-        assert!(warnings.is_empty());
-    }
-
-    #[tokio::test]
-    async fn check_outdated_continues_on_network_failure() {
-        let (mut installer, mock_server, _tmp) = test_installer().await;
-
-        {
-            let tx = installer.db.transaction().unwrap();
-            tx.record_install("good", "1.0.0", "old_sha").unwrap();
-            tx.record_install("bad", "1.0.0", "old_sha").unwrap();
-            tx.commit().unwrap();
-        }
-
-        let bulk = format!("[{}]", formula_json("good", "2.0.0", "new_sha"));
-        Mock::given(method("GET"))
-            .and(path("/formula.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(bulk))
-            .mount(&mock_server)
-            .await;
-
-        Mock::given(method("GET"))
-            .and(path("/formula/bad.json"))
-            .respond_with(ResponseTemplate::new(500))
-            .mount(&mock_server)
-            .await;
-
-        let (outdated, warnings) = installer.check_outdated().await.unwrap();
-        assert_eq!(outdated.len(), 1);
-        assert_eq!(outdated[0].name, "good");
-        assert_eq!(warnings.len(), 1);
-        assert!(warnings[0].contains("bad"));
-    }
-
-    #[tokio::test]
-    async fn check_outdated_warns_on_missing_bottle() {
-        let (mut installer, mock_server, _tmp) = test_installer().await;
-
-        {
-            let tx = installer.db.transaction().unwrap();
-            tx.record_install("nobottle", "1.0.0", "old_sha").unwrap();
-            tx.commit().unwrap();
-        }
-
-        let bulk = r#"[{
-            "name": "nobottle",
-            "versions": { "stable": "2.0.0" },
-            "dependencies": [],
-            "bottle": { "stable": { "files": {} } }
-        }]"#;
-
-        Mock::given(method("GET"))
-            .and(path("/formula.json"))
-            .respond_with(ResponseTemplate::new(200).set_body_string(bulk))
-            .mount(&mock_server)
-            .await;
-
-        let (outdated, warnings) = installer.check_outdated().await.unwrap();
-        assert!(outdated.is_empty());
-        assert_eq!(warnings.len(), 1);
-        assert!(warnings[0].contains("nobottle"));
-    }
-
-    #[test]
-    fn stage_raw_cask_binary_copies_and_marks_executable() {
-        let tmp = TempDir::new().unwrap();
-        let blob_path = tmp.path().join("claude");
-        fs::write(&blob_path, b"#!/bin/sh\necho hello").unwrap();
-
-        let keg_path = tmp.path().join("keg");
-        let cask = crate::installer::cask::ResolvedCask {
-            install_name: "cask:claude-code".to_string(),
-            token: "claude-code".to_string(),
-            version: "1.0.0".to_string(),
-            url: "https://example.com/claude".to_string(),
-            sha256: "aaa".to_string(),
-            binaries: vec![crate::installer::cask::CaskBinary {
-                source: "claude".to_string(),
-                target: "claude".to_string(),
-            }],
-        };
-
-        stage_raw_cask_binary(&blob_path, &keg_path, &cask).unwrap();
-
-        let target = keg_path.join("bin/claude");
-        assert!(target.exists());
-        assert_eq!(
-            fs::read_to_string(&target).unwrap(),
-            "#!/bin/sh\necho hello"
-        );
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            let mode = fs::metadata(&target).unwrap().permissions().mode();
-            assert_eq!(mode & 0o755, 0o755);
-        }
-    }
-
-    #[test]
-    fn stage_raw_cask_binary_rejects_multiple_binaries() {
-        let tmp = TempDir::new().unwrap();
-        let blob_path = tmp.path().join("blob");
-        fs::write(&blob_path, b"data").unwrap();
-
-        let keg_path = tmp.path().join("keg");
-        let cask = crate::installer::cask::ResolvedCask {
-            install_name: "cask:multi".to_string(),
-            token: "multi".to_string(),
-            version: "1.0.0".to_string(),
-            url: "https://example.com/multi".to_string(),
-            sha256: "bbb".to_string(),
-            binaries: vec![
-                crate::installer::cask::CaskBinary {
-                    source: "a".to_string(),
-                    target: "a".to_string(),
-                },
-                crate::installer::cask::CaskBinary {
-                    source: "b".to_string(),
-                    target: "b".to_string(),
-                },
-            ],
-        };
-
-        let err = stage_raw_cask_binary(&blob_path, &keg_path, &cask).unwrap_err();
-        assert!(err.to_string().contains("raw binary"));
-    }
-}
